Made review changes
